ch06: implement sequences
Took a little extra care in `ArraySeq` to ensure it doesn't modify the
array - no `pop` or `unshift` to pull things out. I considered copying
the list first, but indexing in seemed fine.

Kinda wanted my interface to be just a function; you call it to get the
next value, and if it throws a sigil value, then the iteration is done.